updated error message & rename handlers to core_handlers
